[Blink] Reorder DrawingBuffer::FinishPrepareTransferableResourceGpu code
This CL moves the population of the TransferableResource in
DrawingBuffer::FinishPrepareTransferableResourceGpu() to the end of the
method. There are no state dependencies between this population and the
following code in the method, so the reorder is a functional no-op. This
paves the way for splitting this method into two in a followup:
* ExportColorBuffer(), which will have the body of the current method
  but *not* take in or populate a TransferableResource
* FinishPrepareTransferableResourceGpu(), which will wrap the above and
  populate the passed-in TransferableResource

That splitting will in turn remove the need for export from
DrawingBuffer to canvas/static bitmap to indirect through
TransferableResource in further followup.

Bug: 352263194
Change-Id: I5d3cdc1c737b9b93732bccda1324606f498925d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6132820
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401785}