[Passwords] Add UMA for masked and not masked fields.
This CL adds two UMA histograms:
1) PasswordManager.Parsing.PasswordField.IsMasked - it records
if the field that the model predicted to be a password field
is masked (has HTML type = password). It is needed to evaluate
how big the problem of passwords in cleartext is (because
historically we ignore them, and maybe we shouldn't).

2) PasswordManager.Parsing.UnrelatedFields.AnyFieldIsMasked -
it records if any of the fields predicted to not be a part of
a password flow is masked. It is needed to evaluate how often
the fields that we usually supported for filling and saving
just because of the field HTML type, are getting ignored by the
password manager with the model predictions.

Both metrics are recorded only when the form is parsed for
filling, to avoid recording data on every user keystroke during
the parsing in the saving mode.

In the follow-up CLs I'll add a UKM mirror for this metrics
to allow investigation of the most popular domains in each
category.

Bug: 371933424, 312195966
Change-Id: I6f6b833f068571560c4646d4d5966ba49194438f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6138591
Commit-Queue: Maria Kazinova <kazinova@google.com>
Reviewed-by: Viktor Semeniuk <vsemeniuk@google.com>
Cr-Commit-Position: refs/heads/main@{#1401877}