Prevent tab icon animations for instant loads
Previously, tab icons would be animated for all loads, even for
loads that don't need a UI[1], which are same-document navigations
that are typically instant. This creates a potential
flicker effect for near-instant loads.

The effect is currently not visible because BrowserView drives the
loading animation using a 30ms repeating timer, which is too slow to
start the animation before the contents are done loading. However, this
animation will soon be driven by the compositor, which is fast enough
to capture these state changes and displays the flicker.

This CL updates the TabIconViewModel::ShouldTabIconViewAnimate and
TabStripModel::TabsAreLoading (renamed) implementations use WebContents::ShouldShowLoadingUI instead of the older
WebContents::Loading method, to prevent animating instant-loads.

[1]https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/loading_state.h;l=16;drc=cc14db140cbb0fd068bfd271971307fee9a1721b

Bug: 40816290
Change-Id: I4e1e6b0209ab96887bdc8ff4be7ca7f8027c0445
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6132543
Reviewed-by: Shibalik Mohapatra <shibalik@chromium.org>
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401888}