Update WebXR Lighting Estimation Test page
This fixes two errors observed with the lighting estimation test page.

The first is that there were a lot of WebGL errors about rendering
attempting to happen outside of a session rAF. This appears to be due
to the fact that the `animationFrame` callback was in fact being set
on both the XR and non-XR renderer. The fix for this is to only set it
on the XR renderer. This is also speculatively a fix for an inconsistent
repro where the sphere never updates it's color from black.

The second fix is that the reflection map did not appear to be getting
applied. The fix for this requires both updating the `needsPMREMUpdate`
value on the texture and updating the version of THREE.js being used, as
it appears there may have been a bug in the version currently used.

Bug: 380119222
Change-Id: I9465f85a82046e05b5dd272e63025fd5221ea20d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6137999
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Brandon Jones <bajones@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401891}