Use box layout for WindowMiniViewHeaderView.
Box layout is ~10x faster than flex layout in this use case.
For most real-world cases, the difference is probably imperceptible,
but the tast test in the bug surfaced this difference.

Video recordings of this tast test before and after this change
are below.
Before:
http://screencast/cast/NDUyMjE2NTc1OTExNTI2NHw3M2FhYTQ1Ny0zYw
After:
http://screencast/cast/NTY2MTc5NTM2Nzc4MDM1Mnw5OTBkMDg3YS1mNA

      Ran ui.SplitViewResizePerf tast test
      Ran with --ui-show-paint-rects to ensure there's not overpainting.

Bug: b:385148375
Test: Tested both short and long window labels.
Change-Id: Ie1417bb1a71b6b6589b80eab6aec3a63fe6e9f48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6137996
Reviewed-by: Mitsuru Oshima <oshima@chromium.org>
Commit-Queue: Eric Sum <esum@google.com>
Cr-Commit-Position: refs/heads/main@{#1401919}