Avoid fling when trying to hide top controls in tests
The top controls are not immediately scrollable and, when a fling is
ongoing, further flings also don't have the chance to scroll the top
controls.

Thus, avoiding a fling on the first drag reduces flakiness of tests that
need to hide the top controls.

Tested in:
- https://ci.chromium.org/ui/inv/u-acondor-2025-01-03-15-22-00-e11450e43cd885cf
- https://ci.chromium.org/ui/inv/u-acondor-2025-01-03-15-40-00-e22a4c89cdeae201
- https://ci.chromium.org/ui/inv/u-acondor-2025-01-03-15-46-00-1912d39c8c9cb58d

Fixed: 385189006
Change-Id: I61497fe25c82aaf5f18edb51302c789999511bd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6141873
Reviewed-by: Michael Thiessen <mthiesse@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401926}