Update GetAnchorView to accept ActionIds instead of PageActionIconType
This CL is a refactor with no intended changes in behaviour.

Previously, GetAnchorView only accepted PageActionIconType, which was
used to indirectly retrieve the corresponding ActionId, which is then
used to lookup the pinned action view.

Now, the ActionId is provided directly. This is required to support the
Page Actions framework migration, where PageActionIconType will
ultimately be removed and page actions will be identified by their
ActionIds.

Also, the method now can be used for cases where an action is pinnable
but doesn't have a page action, removing duplicated logic.

Note: there are multiple callsites that were providing
PageActionIconTypes for page actions without action IDs. These are
replaced with std::nullopt for now (no change in behaviour), and will be
updated as appropriate action IDs are added during the migration.

Bug: 386362832
Change-Id: I539d1525a99c43a4aa5980167e004597eda4ba8a
Fixed: 386362832
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6114427
Commit-Queue: Kaan Alsan <alsan@chromium.org>
Reviewed-by: Caroline Rising <corising@chromium.org>
Reviewed-by: David Pennington <dpenning@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401938}