Render glic context menu hotkey for all cases on MacOS
When kGlicStatusIconOpenMenuWithSecondaryClick is false, the context
menu is created before the hotkey is set. It is also possible that
the hotkey will be updated after the context menu is created. For that
reason we need a new method that will refresh the context menu.

Bug: 386803748
Change-Id: I111ed23bcd5b466102ec42eb89597266102a4997
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6138876
Reviewed-by: Eshwar Stalin <estalin@chromium.org>
Commit-Queue: Alison Gale <agale@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401953}