Add shortcut to toggle camera access (behind flag)
This CL creates a new shortcut that toggles the setting "Camera access"
in the "Privacy controls" section of ChromeOS Settings. When the setting
is enabled or disabled, a toast message is displayed.

Example of toast message:
https://screenshot.googleplex.com/4FVYuo4rpArdez6

There is a special case where camera access may be force disabled, due
to an admin remoting into a user's ChromeOS instance. In this case,
the shortcut has no effect, and a toast message is shown indicating
that the setting cannot be changed.

Bug: b/383772675
Change-Id: I64fa7e8e0de2427e23f4bae368c64a9166c0ec1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6092031
Commit-Queue: Camden Bickel <cambickel@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Alex Gough <ajgo@chromium.org>
Reviewed-by: David Padlipsky <dpad@google.com>
Reviewed-by: Jiaming Cheng <jiamingc@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1401983}