[Autofill] Refactor internals of BAM::OnFormSubmittedImpl() (2/2)
BAM::OnFormSubmittedImpl() invokes Autofill's form import,
vote uploading, and metrics logging in an `if` and an `else`
branch.

To reduce code duplication, this CL pulls the lambda from
the `if` branch upwards and also calls it in the `else` branch.

This is in preparation of crrev.com/c/6142894.

This CL is a refactoring.

Bug: 374086145
Change-Id: I3eab174bb49cdca46839897c968bab64406f2e81
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6142715
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Dominic Battré <battre@chromium.org>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1402114}