[CCT] Record histogram for URL scheme on context menu
Adds a histogram for URL scheme the context menu is brought up for.
This is done to check the volume of data scheme against the others,
and hide 'Preview page' item in the menu if the volume is low enough.

Low-Coverage-Reason: Class already had a low base coverage
Bug: 362545037
Change-Id: I90e8013bfc7c6749009f0ffed5c33962891f287b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6142553
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Cr-Commit-Position: refs/heads/main@{#1402199}