feedback: Add ability to specify custom image MIME type
The Scanner project will use JPEG images, not PNG images, but there is
currently no way of specifying the MIME type of the image (it is assumed
to be PNG).

Add a `image_mime_type_` string member to `FeedbackCommon`. To preserve
compatibility with existing code, it is assumed that if it is not
specified, the image provided is a PNG.

Bug: b:367882164
Change-Id: I6a6a2100c6098be3c69a8e9494dd8cab3c7bda28
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6091255
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Xiangdong Kong <xiangdongkong@google.com>
Commit-Queue: Michael Cui <mlcui@google.com>
Cr-Commit-Position: refs/heads/main@{#1402258}