MediaSession: Add WebAudio players as ambient players
When the AudioFocusManager ducks a MediaSessionImpl, it lowers the
volume of its underlying players. This does not currently lower the
volume of WebAudio playback, as AudioContexts are not registered to
the MediaSessionImpl as media players.

WebAudio has not been registered before as we have not wanted
WebAudio playback to affect audio focus. However, there's an "ambient"
audio focus type that accomplishes just that (currently unused by
MediaSessionImpl). This CL adds "ambient" players to the
MediaSessionImpl which will request ambient focus if there are no other
players requesting a stronger type of audio focus, and makes
AudioContexts register as ambient players. This allows the
AudioFocusManager to know about WebAudio playback so it can force it
to duck when needed.

On the AudioContext side, it ducks the output volume by scaling the
output audio bus according to the volume multiplier given by the
browser side.

Bug: 382316461
Change-Id: Iffdb572d8c999177b35e17c76717460665bcfbbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6098352
Reviewed-by: Hongchan Choi <hongchan@chromium.org>
Commit-Queue: Tommy Steimel <steimel@chromium.org>
Reviewed-by: Mark Foltz <mfoltz@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402262}