[views-ax] Remove the "GetAccessibleNodeData()" from FindBarView
This CL removes the setting of the cached role and
GetAccessibleNode() call which was added as a temporary
fix in find_bar_view.cc. It also sets the cached name
in GetViewAccessibility().AnnouncePolitely() as part of
push-based architecture.


This CL is part of the ViewsAX project:
https://docs.google.com/document/d/1Ku7HOyDsiZem1yaV6ccZ-tz3lO2XR2NEcm8HjR6d-VY/edit#heading=h.ke1u3utej413

Bug: 325137417
Change-Id: Ife1378d053089c9253322fd6d649332ffabecd6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6124815
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Dileep Maurya <dileepmaurya@microsoft.com>
Reviewed-by: Divyansh Mangal <dmangal@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1402268}