[Signin] Check auth errors in AccountManagedStatusFinder
If a timeout is provided to AccountManagedStatusFinder - consider
persistent auth errors as the timeout outcome. The assumption being that
time-sensitive code paths (which are mostly in the UI) shouldn't wait
for user to resolve the persistent auth error.

Bug: 348373729
Change-Id: Id8dcdf9b4e488966583dcd8758eb5c6559106078
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6120421
Reviewed-by: David Roger <droger@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402301}