[Blink] Separate out DrawingBuffer code for exporting ClientSI
DrawingBuffer has a FinishPrepareTransferableResourceGpu() method in
order to be able to send WebGL-rendered content to the compositor.
However, other callsites use this API unnecessarily simply due to
convenience even though they don't actually need a TransferableResource.

This CL separates out the meat of this method into a new
ExportSharedImageFromBackBuffer() method.
FinishPrepareTransferableResourceGpu() now invokes that method and uses
the ClientSI returned to populate the TransferableResource. The one
piece of state being used to populate the TransferableResource that
*doesn't* come from the ClientSI itself is the SyncToken. We pass the
SyncToken as a reference to ExportSharedImageFromBackBuffer() and have
that method populate it with the correct value for callers to use.

Followup CLs will change the callsites mentioned above to invoke
ExportSharedImageFromBackBuffer() directly.

Bug: 352263194
Change-Id: Ia0f8bba633e7222195b0cd51e3b77a41ed0a4008
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6141833
Reviewed-by: Saifuddin Hitawala <hitawala@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402341}