Implement "Show Notification" intent
When a warning notification is displayed, two actions are presented:
"Unsubscribe" and "Show Notification". This CL implements
the "Show Notification" intent. Once the event is received,
we will retrieve the original notification contents that was
stored as backup, and restore the notification.

Bug: 382475678
Change-Id: Ib518139bc7c22ab2f083c7e6244adb9d0fd94763
Low-Coverage-Reason: COVERAGE_UNDERREPORTED
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6107293
Commit-Queue: Richard Chen <richche@google.com>
Reviewed-by: Sarah Krakowiak <skrakowi@chromium.org>
Reviewed-by: Peter Beverloo <peter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402424}