infra: Fix GN arg mirroring of mac11 trybot
https://ci.chromium.org/ui/p/chromium/builders/try/mac_chromium_11.0_rel_ng

In individual builds, that builder generally has a lot of web test
failures with error messages like:
"No video codecs in common."
and
"FFmpegDemuxer: no supported streams"

Its CI bot doesn't have such errors. This could be due to a difference
in ffmpeg and codec related GN args. So this fixes that mismatch in
the hopes of resolving those errors. In fact, web tests seems much
healthier on that bot with this CL:
https://ci.chromium.org/ui/p/chromium/builders/try/mac_chromium_11.0_rel_ng/1643/infra
https://ci.chromium.org/ui/p/chromium/builders/try/mac_chromium_11.0_rel_ng/1644/infra
https://ci.chromium.org/ui/p/chromium/builders/try/mac_chromium_11.0_rel_ng/1645/infra

Bug: 380911420
Change-Id: I8b23be6c5f538263829f0a2a71de3cf5c413cad7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6127615
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Yue She <yueshe@google.com>
Cr-Commit-Position: refs/heads/main@{#1402465}