[FPF] Add UMAs for refresh heuristic exceptions
In previous CLs (5962525, 5966957), we have implemented a feature flag
param, `refresh_heuristic_exception_threshold`, which, when nonzero,
causes adding a FPF exception when a user refreshes a site multiple
times in the same WebContents (tab). These exceptions are stored in
a profile pref.

This CL adds UMAs to track when exceptions are added, when they are
applied, and the latency of adding and checking for exceptions.

Bug: 372669423
Change-Id: Ic006c4298baa5e6741e2d1edb7bf18bea4a25de2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6011271
Reviewed-by: Patricia Rock <trishalfonso@google.com>
Reviewed-by: Zainab Rizvi <rizvis@google.com>
Commit-Queue: Adam Abeshouse <aabeshouse@google.com>
Cr-Commit-Position: refs/heads/main@{#1402486}