[spark][omnibox] Clean up feature config
No behavior change.

It was confusing to have the config be stateful. This CL removes the
stateful fields and just recomputes the trivial "is valid" logic each
time its asked. Now it looks more like the typical omnibox config, a
struct with a bunch of fields that 1:1 map to finch params.

It was confusing to have the redundant `Init()` methods. All the
constructor would do is call `Init()`. This CL simply moves the code
from `Init()` into the constructor to remove the `Init()` methods.

Bug: 379710012
Change-Id: I79d3cfbe711f117318eec2aec19c8346b6faedaa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6149339
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1402611}