[Protected Audiences] Make KVv2 browser tests set coordinator keys.
This CL migrates the InterestGroupTrustedSignalsKVv2BrowserTest test
fixture to using SetBiddingAndAuctionServerKeys() to directly set the
coordinator keys, instead of mocking out the response to fetching the
keys over the network.

This simplifies the tests a little. Also, there can only be one
URLLoaderInterceptor globally, and this CL removes that, so tests can
use a URLMonitor, which sites on top of URLLoaderInterceptor, to probe
requests.

Bug: None
Change-Id: I25c778a114eaac7e6ded1113420e39d00ca099db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6111304
Commit-Queue: mmenke <mmenke@chromium.org>
Reviewed-by: Russ Hamilton <behamilton@google.com>
Cr-Commit-Position: refs/heads/main@{#1402649}