Move presentation attribute style into ShareableElementData.
Due to Oilpan member compression, there's room in ShareableElementData
for this pointer now, so we can just as well make it shareable.
This allows us to save some work in by just reusing the style when we
have a shared entry (similar to what we do with inline styles).

Memory usage change is a bit complex. On the positive side,
we need to use UniqueElementData less often (more sharing),
and it's also smaller when we do need it (better packing).
But it also requires us to have the tag name as part of the
cache key, which will reduce cache hit rate somewhat and thus
use more memory in some cases. (It also makes each cache entry
slightly larger.) I tried taking out the tag name if there were
no presentation attributes, but it seemed to be perf-negative;
in general, the performance profile of this seems to be somewhat
sensitive to small changes.

Style perftest (Zen 3, LTO but no PGO):

  Initial style (µs)     Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                5479      5484   -0.1%  [ -0.3%,  +0.1%]
  Encyclopedia            45815     45388   +0.9%  [ +0.6%,  +1.3%]
  Extension               52865     52443   +0.8%  [ +0.6%,  +1.0%]
  News                    22796     22404   +1.7%  [ +1.6%,  +1.9%]
  Search                   7876      7794   +1.1%  [ +0.7%,  +1.3%]
  Social1                 14202     14055   +1.0%  [ +0.8%,  +1.3%]
  Social2                  9261      9180   +0.9%  [ +0.7%,  +1.1%]
  Sports                  26600     26693   -0.3%  [ -0.8%,  +0.1%]
  Video                   18050     17867   +1.0%  [ +0.8%,  +1.3%]
  Geometric mean                            +0.8%  [ +0.7%,  +0.9%]

  Recalc style (µs)      Before     After    Perf      95% CI (BCa)
  =================== ========= ========= ======= =================
  ECommerce                3005      2972   +1.1%  [ +0.8%,  +1.4%]
  Encyclopedia            35848     35316   +1.5%  [ +1.2%,  +1.7%]
  Extension               41366     41268   +0.2%  [ -0.0%,  +0.5%]
  News                    11643     11535   +0.9%  [ +0.7%,  +1.2%]
  Search                   2979      2963   +0.6%  [ +0.2%,  +0.9%]
  Social1                  6940      6925   +0.2%  [ -0.0%,  +0.5%]
  Social2                  4951      4899   +1.1%  [ +0.8%,  +1.3%]
  Sports                  14264     14086   +1.3%  [ +1.0%,  +1.5%]
  Video                    7798      7749   +0.6%  [ +0.4%,  +0.9%]
  Geometric mean                            +0.8%  [ +0.7%,  +1.0%]

MotionMark (M1 Pinpoint, significant results at 99% CI only, higher is
better):

  Suits                      [ +0.0%,  +1.0%]
  Leaves                     [ +0.1%,  +0.8%]
  Multiply                   [ +0.5%,  +1.0%]
  motionmark_ramp_composite  [ +0.0%,  +0.6%]

Speedometer3 (M1 Pinpoint, same, lower is better except for Score):

  Editor-TipTap                               [ -0.4%,  -0.0%]
  TodoMVC-Preact-Complex-DOM                  [ -2.8%,  -0.1%]
  TodoMVC-JavaScript-ES6-Webpack-Complex-DOM  [ -0.9%,  -0.2%]
  React-Stockcharts-SVG                       [ -1.0%,  -0.3%]
  Score                                       [ +0.1%,  +0.4%]

Change-Id: I2a53d4bfcd3804331d4a75f7ba98b1dda726444a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6105791
Reviewed-by: Mason Freed <masonf@chromium.org>
Commit-Queue: Steinar H Gunderson <sesse@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402655}