[discard] Ensure tabs remain eligible for successive discard operations
This CL updates PageDiscardingHelper to reset DiscardAttemptMarkers
for nodes following a main frame navigation.

This is necessary to ensure that a given tab remains eligible for
discarding in the specific case where:
  1. A discard attempt was previously made on the tab
  2. The tab was later navigated / reloaded and its content is
     refreshed

The current behavior assumes that the DiscardAttemptMarker is
destroyed following a discard operation, as the target WebContents
/ PageNode is replaced.

However with the new implementation the discarded tab's WebContents
is retained. This makes the tab inelibile for any future discard
operations, even after it has been re-activated  / reloaded.

See bug for additional context.

Bug: 386801193
Change-Id: Id675d8a349d95d7f4ca34620c5ba57d254634210
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6132978
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Commit-Queue: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402667}