[BCIV] Fix status indicator animations
Status indicator animations change the height of the controls and
viewport, so the browser needs to update the offset to reflect the new resting position of the browser controls.

When the controls are scrolled off the screen, setting the offset with the content offset is incorrect, because viz will receive a frame with the scroll offset applied, and will apply them again via the offset tag values.

This CL updates the browser with the correct offsets (using the height, not including any scroll offsets) when the min height of the controls change, so that the offset tag values in viz will be correctly applied. This fix isn't perfect, but should be good enough for stable experimentation.

See bug for videos of before/after.

Bug: 387913827
Change-Id: I7b24852facd89f03d8efaeb1108e5f8f3c0a1de7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6149868
Reviewed-by: Wenyu Fu <wenyufu@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1402684}